requirements.txt
aio-pika==9.1.4
celery
fastapi
uvicorn
flower
tiktoken
httpx==0.24.1
loguru==0.7.1
pymongo==4.5.0
pydantic-settings==2.0.2
python-dotenv==1.0.0
neo4j==5.11
openai==0.27.4
langchain==0.0.351
jose== 1.0.0
jose== 1.0.0
passlib==1.7.4
jwt==1.3.1
python-jose==3.3.0
python-decouple==3.8
python-multipart==0.0.6
bcrypt==4.1.2
flower==2.0.1
beautifulsoup4==4.12.2
# dev
pytest==7.4.1
pytest-asyncio==0.21.1